pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.6.5",
    "bcrypt>=5.0.0",
    "beautifulsoup4>=4.13.4",
    "flask>=3.1.1",
    "func-timeout>=4.3.5",
    "google-cloud-vision>=3.10.2",
    "lxml>=6.0.0",
    "pillow>=12.0.0",
    "requests>=2.32.4",
    "selenium>=4.39.0",
]
